Port "file_util: Minor changes to ScanDirectoryTree() and ForeachDirectoryEntry()" from yuzu

+++ src/citra_qt/game_list.cpp
@@ -630,7 +630,7 @@ void GameList::RefreshGameDirectory() {
 
 void GameListWorker::AddFstEntriesToGameList(const std::string& dir_path, unsigned int recursion,
                                              GameListDir* parent_dir) {
-    const auto callback = [this, recursion, parent_dir](unsigned* num_entries_out,
+    const auto callback = [this, recursion, parent_dir](u64* num_entries_out,
                                                         const std::string& directory,
                                                         const std::string& virtual_name) -> bool {
         std::string physical_name = directory + DIR_SEP + virtual_name;
